[CARBONDATA-1218] In case of data-load failure the BadRecordsLogger.badRecordEntry map holding the task Status is not removing the task Entry
Problem
For GLOBAL_SORT scope option in case of data-load failure the BadRecordsLogger.badRecordEntry map holding the task Status is not removing the task Entry.
Because of this the next load is getting failed even though the data being loaded has no bad records.

Solution
The map entry must be removed after load completion either success or fail.
Refactored the Bad record logger.

This closes #1082
